[feat/#8] form 의 상수, 타입 선언 분리
