+++ src/types/form.ts
@@ -1,12 +1,39 @@
-import { FormKeys, IconKeys, Position } from "@/src/constants/form";
+import { fieldPosition } from "../constants/form";
 import ButtonProps from "@/src/types/button";
 import { InputHTMLAttributes, ReactNode, TextareaHTMLAttributes } from "react";
 import { FieldValues } from "react-hook-form";
+
+export type Position = "left" | "right";
+
+// export type IconKeys = keyof typeof icons; 이렇게 하고 싶은데 이렇게 하면 자동완성이 안됨
+export type IconKeys = "password" | "confirmPassword" | "calendar" | "search";
+export type IconType = {
+  icon: (isVisible?: boolean) => ReactNode;
+  onClick: () => void;
+  position: Position;
+  padding?: string;
+  margin?: string;
+};
+
+export type InputKeys =
+  | "password"
+  | "confirmPassword"
+  | "nickname"
+  | "email"
+  | "default"
+  | "search";
+
+export type FieldKeys = "authPage" | "expPage" | "search" | "default";
+export type FormKeys = InputKeys | FieldKeys;
 
 export interface BaseProps {
   children?: ReactNode;
   className?: string;
   resetOnSubmit?: boolean;
+}
+
+export interface FieldStyles {
+  style: (typeof fieldPosition)[keyof typeof fieldPosition];
 }
 
 export interface FormProps<T extends FieldValues> extends BaseProps {
@@ -26,6 +53,12 @@ export interface LabelProps
 export interface IconProps extends BaseProps {
   onClick?: () => void;
   position: Position;
+}
+
+export interface ButtonStyles {
+  style?: ButtonProps;
+  tailwindStyle?: string;
+  fieldStyle?: string;
 }
 
 export interface ErrorProps extends BaseProps {}
